RSPY-160: Splitted endpoints test file into multiple test files
